Remove requirement node-inspector
Node-inspector is deprecated as functionality is built-in to node itself now. Keeping the dependency causes many problems similar to #157 for node-inspector and v8-debug packages. After removing node-inspector, it is possible to install and run tribeca fully functionally on node-js carbon and npm 5.5.1 on win7 x64, which was not possible before.